Fix float casting

+++ app/src/main/java/in/testpress/testpress/ui/adapters/ContentsCarouselAdapter.java
@@ -124,8 +124,8 @@ public class ContentsCarouselAdapter extends RecyclerView.Adapter<ContentsCarous
 
             if (attempt.getUserVideoId() != null) {
                 VideoAttempt userVideo = response.getUserVideoHashMap().get(attempt.getUserVideoId());
-                float lastPosition = (float) Double.parseDouble(userVideo.getLastPosition());
-                float totalDuration = (float) Double.parseDouble(userVideo.getRawVideoContent().getDuration());
+                float lastPosition = Float.parseFloat(userVideo.getLastPosition());
+                float totalDuration = Float.parseFloat(userVideo.getRawVideoContent().getDuration());
                 float watchPercentage = (lastPosition / totalDuration) * 100;
                 holder.videoProgress.setProgress((int) watchPercentage);
                 holder.videoProgressLayout.setVisibility(View.VISIBLE);
